+++ pkg/roles/tftp/read.go
@@ -1,0 +1,7 @@
+package tftp
+
+import "io"
+
+func (r *Role) readHandler(filename string, rf io.ReaderFrom) error {
+	return nil
+}
